refactor: レガシーサービスを整理しTypeScriptエラーを解決
- レガシーのzennService.tsを削除（モックデータとキャッシュ管理を含む）
- convexZennService.tsをzennService.tsにリネーム
- ConvexZennServiceクラスをZennServiceにリネーム
- ZennTrends.tsでのインポートとクラス使用を更新
- 未使用のZennPostインポートを削除
- CLAUDE.mdに設計原則セクションを追加（SOLID, YAGNI, KISS, DRY）

これによりCIで発生していたTypeScriptエラーが解決され、
コードベースからレガシーコードが除去されました。
